slight bug in tws-sort
It was sorting entries based on total times per category correctly,
however it was not using all of the entries in the database to
calculate per-category totals. Instead, it was calculating totals
based on just those entries that were passed in as an argument to
tws-sort. This has been corrected.